everything up to Applications & Examples should now be in a much more streamlined version, with all the notation updated/clarified, and previous issues now corrected.
needs a quick pass at App & Ex again, then a solid clean up of Discussion. In App & Ex, I need to revisit the final example as it really could be done in various ways. The question now is which is the best for this paper vs. which is the best more generally